refactor(Web): move more vebose service configuration to extension methods
Move app service registration into extension methods to provide simpler calls in program file. Keeps separate application-specific service registation and those
that are more platform/implementation specific, e.g., Razor pages, server-side, etc.
