+++ NjordinSight.Web/Program.cs
@@ -38,47 +38,9 @@ var config = BuildConfiguration(logger, databaseProvider == "SQL_SERVER");
 builder.Services.AddSingleton(implementationInstance: logger);
 builder.Services.AddSingleton(implementationInstance: config);
 
-builder.Services.AddSingleton(ISvgHelper.Create());
-
 #endregion
 
-#region Add database contexts
-
-// Add database service.
-builder.Services.AddDbContextFactory<FinanceDbContext>(options =>
-{
-
-    if (string.IsNullOrEmpty(databaseProvider) && builder.Environment.IsDevelopment())
-        options.UseInMemoryDatabase(app_db_name);
-
-    else if (databaseProvider == "SQL_SERVER")
-        options.UseSqlServer(
-            connectionString: $"Name=ConnectionStrings:{app_db_name}",
-            sqlServerOptionsAction: x => x.MigrationsAssembly("NjordinSight.EntityMigration"));
-    else
-        throw new NotSupportedException();
-});
-
-
-// Add identity management database service
-builder.Services.AddDbContext<IdentityDbContext>(options =>
-{
-    if (string.IsNullOrEmpty(databaseProvider) && builder.Environment.IsDevelopment())
-        options.UseInMemoryDatabase(identity_db_name);
-
-    else if (databaseProvider == "SQL_SERVER")
-        options.UseSqlServer(
-            connectionString: $"Name=ConnectionStrings:{identity_db_name}");
-    else
-        throw new NotSupportedException();
-});
-
-if(builder.Environment.IsDevelopment() && string.IsNullOrEmpty(databaseProvider))
-{
-    SeedInMemoryDatabase();
-}
-
-#endregion
+builder.ConfigureDbContextFactories(databaseProvider);
 
 #region Authentication configuration
 
@@ -92,27 +54,12 @@ builder.Services.AddDatabaseDeveloperPageExceptionFilter();
 
 #endregion
 
-
-#region Data-access service configuration
-
-// Add metadata, search, and message services.
-builder.Services.AddSingleton<IExpressionBuilder, ExpressionBuilder>();
-builder.Services.AddSingleton<IModelMetadataService, ModelMetadataService>();
-builder.Services.AddSingleton(typeof(ITypedMetadataService<>), typeof(TypedMetadataService<>));
-
-builder.Services.AddTransient(typeof(ISearchService<>), typeof(SearchService<>));
-
-builder.Services.AddSingleton<IMessageService, MessageService>();
-builder.Services.AddSingleton<IFinancialCalculator, FinancialCalculator>();
-
 // Register model services and controllers.
 builder.Services.AddModelServices();
 builder.Services.AddModelControllers();
 
-// Add auxiliary business-logic services
+// Add auxiliary business-logic and view services
 builder.Services.AddAuxiliaryServices();
-
-#endregion
 
 builder.Services.AddRazorPages();
 builder.Services.AddServerSideBlazor();
@@ -149,16 +96,6 @@ app.Run();
 
 partial class Program
 {
-    /// <summary>
-    /// Database or database store name for the core application entity model.
-    /// </summary>
-    private const string app_db_name = "NjordWorks";
-
-    /// <summary>
-    /// Database or database store for the web identity management entity model.
-    /// </summary>
-    private const string identity_db_name = "NjordIdentity";
-
     /// <summary>
     /// Builds the application <see cref="ILogger"/> instance.
     /// </summary>
@@ -237,18 +174,5 @@ partial class Program
         return new SerilogLoggerFactory(logger).CreateLogger(nameof(Program));
     }
 
-    /// <summary>
-    /// Recreates the 'NjordWorks' database, but takes no action on the 'NjordIdentity' database.
-    /// Call only once during start-up to seed data for an in-memory data store.
-    /// Only for use in development/demonstration purposes.
-    /// </summary>
-    private static void SeedInMemoryDatabase()
-    {
-        var optionsBuild = new DbContextOptionsBuilder<FinanceDbContext>();
-        optionsBuild.UseInMemoryDatabase(app_db_name);
-            
-        using var context = new FinanceDbContext(optionsBuild.Options);
-
-        context.Database.EnsureCreated();
-    }
+    
 }
